backend/requirements.txt
Django>=5.2.4
djangorestframework>=3.14.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
django-cors-headers>=4.3.0 